Add "discord" id to newsletter component
Currently, we tell students to go to our website, scroll down to the form, and input their info in order to join the Discord.
It would be better if we had a link to just take users directly to the form.

By adding `id="discrod"` to that div in the newsletter component, we can now simply send people the following url: https://hack.ics.uci.edu/#discord
This behaves the same way as https://hack.ics.uci.edu/#newsletter

Although it doesn't change the appearance or functionality of anything else on the site, I will concede that adding an id in like that might be slightly hacky/not best practice.